Most recent collision stuff I'll continue more tomorrow

+++ UpgradePlatformer/UpgradePlatformer/Levels/Level.cs
@@ -59,7 +59,7 @@ namespace UpgradePlatformer.Levels
         {
             List<Tile> Tiles = new List<Tile>();
 
-            Point centerTile = r.Center / new Point(Tile.WINDOW_SIZE / 30);
+            Point centerTile = r.Center / new Point((int)Tile.WINDOW_SIZE / 30);
 
             for (int x = Math.Max(centerTile.X - RedundancySize, 0); x < Math.Min(centerTile.X + RedundancySize, TileWidth); x++)
                 for (int y = Math.Max(centerTile.Y - RedundancySize, 0); y < Math.Min(centerTile.Y + RedundancySize, TileHeight); y++)
